redefine exceptions so that they share the same common supertype
