+++ src/main/java/fi/digitraffic/tis/aws/s3/S3Client.java
@@ -79,12 +79,3 @@ public class S3Client {
     }
 }
 
-class S3ClientException extends RuntimeException {
-    public S3ClientException(String message) {
-        super(message);
-    }
-
-    public S3ClientException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
