+++ src/main/java/fi/digitraffic/tis/http/HttpClientUtilityException.java
@@ -1,0 +1,13 @@
+package fi.digitraffic.tis.http;
+
+import fi.digitraffic.tis.TisException;
+
+class HttpClientUtilityException extends TisException {
+    public HttpClientUtilityException(String message) {
+        super(message);
+    }
+
+    public HttpClientUtilityException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
