+++ src/main/java/fi/digitraffic/tis/vaco/ruleset/model/Category.java
@@ -1,7 +1,7 @@
 package fi.digitraffic.tis.vaco.ruleset.model;
 
 import fi.digitraffic.tis.utilities.model.PersistableEnum;
-import fi.digitraffic.tis.vaco.validation.model.InvalidMappingException;
+import fi.digitraffic.tis.vaco.InvalidMappingException;
 
 public enum Category implements PersistableEnum {
     GENERIC("generic"),
