--- src/main/java/fi/digitraffic/tis/vaco/validation/ValidationProcessException.java
@@ -1,16 +1,0 @@
-package fi.digitraffic.tis.vaco.validation;
-
-import fi.digitraffic.tis.vaco.VacoException;
-
-/**
- * Common wrapping supertype for all non-recoverable exceptions thrown during validation process.
- */
-public class ValidationProcessException extends VacoException {
-    public ValidationProcessException(String message) {
-        super(message);
-    }
-
-    public ValidationProcessException(String message, Exception cause) {
-        super(message, cause);
-    }
-}
