+++ src/main/java/fi/digitraffic/tis/vaco/validation/rules/netex/EnturNetexValidatorRule.java
@@ -7,7 +7,7 @@ import fi.digitraffic.tis.vaco.process.model.PhaseData;
 import fi.digitraffic.tis.vaco.queuehandler.model.Entry;
 import fi.digitraffic.tis.vaco.queuehandler.model.ValidationInput;
 import fi.digitraffic.tis.vaco.ruleset.RulesetRepository;
-import fi.digitraffic.tis.vaco.validation.ValidationProcessException;
+import fi.digitraffic.tis.vaco.validation.RuleExecutionException;
 import fi.digitraffic.tis.vaco.validation.model.FileReferences;
 import fi.digitraffic.tis.vaco.validation.model.ImmutableValidationReport;
 import fi.digitraffic.tis.vaco.validation.model.ValidationReport;
@@ -102,7 +102,7 @@ public class EnturNetexValidatorRule extends ValidatorRule implements Rule {
                     phaseData.phase().id(),
                     rulesetRepository.findByName(RULE_NAME).orElseThrow().id(),
                     message));
-            throw new ValidationProcessException(message, e);
+            throw new RuleExecutionException(message, e);
         }
     }
 
@@ -119,7 +119,7 @@ public class EnturNetexValidatorRule extends ValidatorRule implements Rule {
                     phaseData.phase().id(),
                     rulesetRepository.findByName(RULE_NAME).orElseThrow().id(),
                     message));
-            throw new ValidationProcessException(message, e1);
+            throw new RuleExecutionException(message, e1);
         }
         return zipFile;
     }
@@ -136,7 +136,7 @@ public class EnturNetexValidatorRule extends ValidatorRule implements Rule {
                     phaseData.phase().id(),
                     rulesetRepository.findByName(RULE_NAME).orElseThrow().id(),
                     message));
-            throw new ValidationProcessException(message, e);
+            throw new RuleExecutionException(message, e);
         }
         return bytes;
     }
